LayoutTests/platform/mac/media/media-can-play-webm-expected.txt

Test HTMLMediaElement canPlayType() method for webm media containers.

These tests may be expected to fail if the WebKit port does not support the format.

EXPECTED (video.canPlayType('audio/webm') == 'maybe'), OBSERVED '' FAIL
EXPECTED (video.canPlayType('video/webm') == 'maybe'), OBSERVED '' FAIL
EXPECTED (video.canPlayType('audio/webm; codecs=vorbis') == 'probably'), OBSERVED '' FAIL
EXPECTED (video.canPlayType('video/webm; codecs=vp8,vorbis') == 'probably'), OBSERVED '' FAIL
END OF TEST
